refactor(modules): move angular-hint-modules to subdir
