+++ src/modules/angular-hint-modules/display.js
@@ -1,4 +1,4 @@
-var hintLog = angular.hint = require('./log'),
+var hintLog = angular.hint = require('./../log'),
     MODULE_NAME = 'Modules';
 
 module.exports = function(modules) {
