+++ src/modules/angular-hint-modules/formatMultiLoaded.js
