+++ src/modules/angular-hint-modules/getModule.js
