+++ src/modules/angular-hint-modules/inAttrsOrClasses.js
