+++ src/modules/angular-hint-modules/moduleData.js
