+++ src/modules/angular-hint-modules/ngViewNoNgRoute.js
