+++ src/modules/angular-hint-modules/normalizeAttribute.js
