+++ src/modules/angular-hint-modules/start.js
