+++ src/modules/angular-hint-modules/storeDependencies.js
