+++ src/modules/angular-hint-modules/storeNgAppAndView.js
