+++ src/modules/angular-hint-modules/storeUsedModules.js
